requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.4
djangorestframework==3.6.4
enum34==1.1.6
httpie==0.9.9
idna==2.6
ipython==5.4.1
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
